Remove MANAGE_EXTERNAL_STORAGE permission
We only relied on this permission because file_picker
didn't work correctly. It would start the correct
android intent ACTION_OPEN_DOCUMENT or
ACTION_OPEN_DOCUMENT_TREE but perhaps
it doesn't wait correctly for it to be accepted?
Because it would always result in permissions
errors.
I just made some native kotlin code myself instead
of relying on this broken library.