Add graph peeking
In Settings -> Appearance
now there is a setting
"Peek graph". Enabling this
puts a small line graph
at the top of the Graphs
page. It's the data for the
most recent exercise.

The logic behind this feature
is most people flick over
to the graphs page to see
the line chart of the most
recent exercise they just
completed. So we might
as well have it already there.

Disabled by default because
I'm not sure if it's too
cluttery or if it will cause
performance issues. I'll
turn it on and give it a try
for a while first.