Make exercise drag handles bigger
I didn't do it on the plans page
because that has more info we
want to express. The title of the
plan, all the different exercises.
whereas in StartPlan you just need to
say the gist of the exercise name,
so it's ok to make the button bigger.

Also I kept missing this button with my
fat fingers so hopefully makes the
experience better.