Handle failed migrations
If the app fails to upgrade the database,
the user now can choose to:
1. backup the db
2. delete and exit
3. create a github issue

Lots of errors which I eventually find
could be solved by a complete wipe &
re-creation, so the delete button is nice.
Also i've automatically linked the error
to the new github issue which should
make it super easy for me to fix these.