Add first draft of self updates
The AI slop code seems to be doing a
bunch of outdated permissions requesting
related to external storage in the download
and install updates section (for android).

I wanted to see if this type of feature is easy
to setup and it might be, but not this way. Maybe
i'll look into this again at a later date.