Use streams in start plan page instead of promises
This is the starting point for me removing all
the manual state management from plans.
I think most of this stopped using streams
because I was struggling to figure out the
re-ordering feature of plans, but now that I know
how to do it properly we can just stream
everything and not worry about stale data.